+++ Editor/Package.cs
@@ -1,10 +1,11 @@
 ﻿
 using UnityEditor;
 
-namespace Hananoki.SceneViewTools {
+namespace HananokiEditor.SceneViewTools {
   public static class Package {
     public const string name = "SceneViewTools";
+    public const string nameNicify = "Scene View Tools";
     public const string editorPrefName = "Hananoki.SceneViewTools";
-    public const string version = "1.0.9";
+    public const string version = "1.0.10";
   }
 }